Move surcharge variable into reform per review feedback
- Move vt_h619_income_tax_surcharge class into vt_h619_reform.py
- Remove in_effect check from formula (reform activation handles this)
- Add in_effect check to create_vt_h619_reform factory function
- Update tests to use reform syntax
- Delete standalone variable file
